API: move branching_plan to recovery.py
